requirements.txt
Flask
Flask-SQLAlchemy
flask-migrate
flask-login
Flask-WTF
email-validator
python-slugify